pyproject.toml
[tool.poetry]
name = "api-bot-backend"
version = "0.1.0"
description = ""
authors = ["rathodparmeshwar <rathodparmeshwar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.10"
sqlalchemy = "^2.0.38"
asyncpg = "^0.30.0"
pydantic-settings = "^2.8.1"
uvicorn = "^0.34.0"
jwt = "^1.3.1"
python-jose = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
